requirements.txt
opencv-python>=4.5.3
ultralytics>=8.0.0
Pillow>=8.4.0
numpy>=1.21.2